Don't use unsigned long in the SHA1 code; we want 32-bit variables and unsigned long is 64 bits on 64-bit platforms.  Use unsigned int or u_int32_t instead.  Pointed out by Oleg Makarenko.

+++ linux/mppe/sha1.c
@@ -29,7 +29,7 @@
 
 #include "sha1.h"
 
-static void SHA1_Transform(unsigned long[5], const unsigned char[64]);
+static void SHA1_Transform(unsigned int[5], const unsigned char[64]);
 
 #define rol(value, bits) (((value) << (bits)) | ((value) >> (32 - (bits))))
 
@@ -57,12 +57,12 @@ static void SHA1_Transform(unsigned long[5], const unsigned char[64]);
 /* Hash a single 512-bit block. This is the core of the algorithm. */
 
 static void
-SHA1_Transform(unsigned long state[5], const unsigned char buffer[64])
+SHA1_Transform(unsigned int state[5], const unsigned char buffer[64])
 {
-    unsigned long a, b, c, d, e;
+    unsigned int a, b, c, d, e;
     typedef union {
 	unsigned char c[64];
-	unsigned long l[16];
+	unsigned int l[16];
     } CHAR64LONG16;
     CHAR64LONG16 *block;
 
@@ -156,7 +156,7 @@ SHA1_Update(SHA1_CTX *context, const unsigned char *data, unsigned int len)
 void
 SHA1_Final(unsigned char digest[20], SHA1_CTX *context)
 {
-    unsigned long i, j;
+    unsigned int i, j;
     unsigned char finalcount[8];
 
     for (i = 0; i < 8; i++) {
