+++ linux/mppe/sha1.h
@@ -4,8 +4,8 @@
 #define _SHA1_H
 
 typedef struct {
-    unsigned long state[5];
-    unsigned long count[2];
+    unsigned int state[5];
+    unsigned int count[2];
     unsigned char buffer[64];
 } SHA1_CTX;
